feat: extend list of lspconfig "root_dir" paths with ".git"
Projects in which the `tailwind.config.{js,cjs,mjs,ts}` is not located
in the root (e.g. located in `assets/`) will not have the Tailwind
language server get started properly by `nvim-lspconfig`.

Add `.git` to the paths/filenames, as this is probably a path/filename
that will more frequently lead to identifying the root of the project.